Simulator test host working against incomplete API
